Use class-validator for payload
